Celebriity Anagram tool
Front end almost complete - Need to add event listener to Input to read value when enter key is pressed.

Logic - The anagram is providing unexpected results. Need to make it compare 2 objects that count the occurance of letters in the anagram and the celebrity name and use the % of match.